# src/main/resources/openapi/api-definition.yaml
openapi: 3.0.3
info:
  title: Gestión de cuentas bancarias API
  version: 1.0.0

tags:
  - name: Clientes
    description: Operaciones relacionadas con los clientes.
  - name: Cuentas
    description: Operaciones relacionadas con las cuentas bancarias.

paths:
  /clientes:
    get:
      tags:
        - Clientes
      summary: Obtener todos los clientes con sus cuentas asociadas.
      operationId: getCustomersAndAccounts
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
      responses:
        '200':
          description: Lista paginada de todos los clientes con sus datos personales y las cuentas bancarias asociadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomerAccount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /clientes/mayores-de-edad:
    get:
      tags:
        - Clientes
      summary: Obtener todos los clientes que sean mayores de edad (18 años o más).
      operationId: getAdultCustomers
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
      responses:
        '200':
          description: Lista paginada de todos los clientes mayores de edad (que tienen 18 años o más).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /clientes/con-cuenta-superior-a/{cantidad}:
    get:
      tags:
        - Clientes
      summary: Obtener todos los clientes cuya suma total de todas sus cuentas superan la cantidad indicada.
      operationId: getCustomersWithHigherAmount
      parameters:
        - name: cantidad
          in: path
          required: true
          description: Cantidad por la que filtrar clientes.
          schema:
            type: number
            format: bigdecimal
            minimum: 0
            example: 1000.50
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/SizeQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
      responses:
        '200':
          description: Lista paginada de todos los clientes que tienen una suma total mayor a la indicada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /clientes/{dni}:
    get:
      tags:
        - Clientes
      summary: Obtener el detalle de un cliente por su DNI.
      operationId: getCustomerByDni
      parameters:
        - name: dni
          in: path
          required: true
          description: DNI del cliente a consultar.
          schema:
            type: string
            pattern: '^\d{8}[A-Za-z]$'
            example: '12345678A'
      responses:
        '200':
          description: Detalle del cliente solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /cuentas:
    post:
      tags:
        - Cuentas
      summary: Dar de alta una nueva cuenta bancaria
      description: |
        Servicio para dar de alta una nueva cuenta bancaria. Si el cliente existe, se le asocia la cuenta;
        si no existe se crea el nuevo cliente con la cuenta asociada.
      operationId: createBankAccountForCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankAccountForCustomerRequest'
      responses:
        '201':
          description: Cuenta creada con el location del detalle y el objeto creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountNoCustomer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /cuentas/{apiIdCuenta}:
    get:
      tags:
        - Cuentas
      summary: Obtener el detalle de una cuenta bancaria por su API ID.
      operationId: getBankAccountByApiId
      parameters:
        - name: apiIdCuenta
          in: path
          required: true
          description: API ID de la cuenta bancaria a consultar.
          schema:
            type: string
            format: uuid
            example: '123e4567-e89b-12d3-a456-426614174000'
      responses:
        '200':
          description: Detalle de la cuenta bancaria solicitada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Cuentas
      summary: Actualizar el saldo total de una cuenta bancaria por su API ID.
      operationId: updateBankAccountTotal
      parameters:
        - name: apiIdCuenta
          in: path
          required: true
          description: API ID de la cuenta bancaria a actualizar.
          schema:
            type: string
            format: uuid
            example: '123e4567-e89b-12d3-a456-426614174000'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                total:
                  type: number
                  format: bigdecimal
                  minimum: 0
                  example: 1500.75
              required:
                - total
      responses:
        '204':
          description: Cuenta bancaria actualizada correctamente. No devuelve contenido.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    PageQueryParam:
      name: page
      in: query
      required: false
      description: Número de página (comenzando desde 0)
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeQueryParam:
      name: size
      in: query
      required: false
      description: Tamaño de la página
      schema:
        type: integer
        minimum: 1
        default: 20
    SortQueryParam:
      name: sort
      in: query
      required: false
      description: Criterio de ordenamiento (formato field,direction - ej. name,asc)
      schema:
        type: string
  schemas:
    Problem:
      $ref: './components/schemas.yaml#/Problem'
    BankAccountNoCustomer:
      $ref: './components/schemas.yaml#/BankAccountNoCustomer'
    BankAccount:
      $ref: './components/schemas.yaml#/BankAccount'
    CustomerAccount:
      $ref: './components/schemas.yaml#/CustomerAccount'
    Customer:
      $ref: './components/schemas.yaml#/Customer'
    PaginationMetadata:
      $ref: './components/schemas.yaml#/PaginationMetadata'
    PaginatedCustomerAccount:
      $ref: './components/schemas.yaml#/PaginatedCustomerAccount'
    PaginatedCustomer:
      $ref: './components/schemas.yaml#/PaginatedCustomer'
    CreateBankAccountForCustomerRequest:
      $ref: './components/schemas.yaml#/CreateBankAccountForCustomerRequest'
  responses:
    BadRequest:
      $ref: './components/responses.yaml#/BadRequest'
    Unauthorized:
      $ref: './components/responses.yaml#/Unauthorized'
    Forbidden:
      $ref: './components/responses.yaml#/Forbidden'
    NotFound:
      $ref: './components/responses.yaml#/NotFound'
    InternalServerError:
      $ref: './components/responses.yaml#/InternalServerError'